fix float64

+++ espnet/nets/batch_beam_search.py
@@ -158,7 +158,7 @@ class BatchBeamSearch(BeamSearch):
             weighted_scores += self.weights[k] * scores[k]
         for k in self.part_scorers:
             weighted_scores[part_ids] += self.weights[k] * part_scores[k]
-        weighted_scores += running_hyps.score.unsqueeze(1)
+        weighted_scores += running_hyps.score.to(dtype=x.dtype).unsqueeze(1)
 
         # TODO(karita): do not use list. use batch instead
         # see also https://github.com/espnet/espnet/pull/1402#discussion_r354561029
